Committed New Framework and Modules
Created a new framework for us to work on. It's called Electron Forge and will be a great starting asset for producing our software.

Here's the URL for the documentation of it:
https://www.electronforge.io/

I also used the README.md to put what resources we are using for this project. You'll see that I added the framework and VS & VSC (Visual Studio and Visual Studio Code, respectively).